Corregidos diversos errores y cambio en el modo de numerar los lors
Se corrigió el script de opencl. Se cambió a otra formula (una antigua)
de numeracion de LORS. Se corrigieron errores en la combinatoria de los
cristales para generar los LORS, y se corrigieron funciones en vector.c
que usaban double en vez de float.

+++ lors_getter/vector.c
@@ -6,9 +6,9 @@
 
 //******************************************************************************
 float module(vector *entrada){
-	float x2 = pow(entrada->x, 2);
-	float y2 = pow(entrada->y, 2);
-	float z2 = pow(entrada->z, 2);
+	float x2 = powf(entrada->x, 2);
+	float y2 = powf(entrada->y, 2);
+	float z2 = powf(entrada->z, 2);
 	return sqrt(x2 + y2 + z2);
 }
